# api/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api_handler.CreateTenantRequest:
    properties:
      description:
        type: string
      id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    required:
    - id
    - name
    type: object
  internal_api_handler.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
...
  internal_api_handler.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  internal_api_handler.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    多租户文件索引服务，支持 PDF、图片、文档等文件的索引、存储和搜索
    认证使用 JWT Token，数据按租户隔离存储在独立的 OpenSearch 索引中。
  title: OpenSearch File API
  version: 1.0.0
paths:
  /admin/tenants:
    get:
      description: 获取所有租户的列表，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 列出租户
      tags:
      - Tenant
    post:
      consumes:
      - application/json
      description: 创建一个新的租户，需要提供唯一的租户 ID 和名称
      parameters:
      - description: 创建租户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建租户
      tags:
      - Tenant
  /admin/tenants/{id}:
    delete:
      description: 软删除租户，租户状态变为 deleted，但数据仍保留
      parameters:
      - description: 租户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除租户（软删除）
      tags:
      - Tenant
    get:
      description: 根据租户 ID 获取详细信息
      parameters:
      - description: 租户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取租户信息
      tags:
      - Tenant
    put:
      consumes:
      - application/json
      description: 更新租户信息，只能更新 name、description 和 metadata 字段
      parameters:
      - description: 租户 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新租户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新租户
      tags:
      - Tenant
  /admin/tenants/{id}/hard:
    delete:
      description: 彻底删除租户，不可恢复
      parameters:
      - description: 租户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 彻底删除租户（硬删除）
      tags:
      - Tenant
...
securityDefinitions:
  BearerAuth:
    description: JWT Token，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  X-Tenant-ID:
    description: 租户 ID
    in: header
    name: X-Tenant-ID
    type: apiKey
swagger: "2.0"